resolved rest of comments left on pr
