+++ src/commands/cluster/config_interfaces/cluster_ref_connect_context.ts
@@ -1,0 +1,8 @@
+/**
+ * SPDX-License-Identifier: Apache-2.0
+ */
+import {type ClusterRefConnectConfigClass} from './cluster_ref_connect_config_class.js';
+
+export interface ClusterRefConnectContext {
+  config: ClusterRefConnectConfigClass;
+}
