+++ src/commands/cluster/configs.ts
@@ -8,7 +8,6 @@ import {ListrInquirerPromptAdapter} from '@listr2/prompt-adapter-inquirer';
 import {confirm as confirmPrompt} from '@inquirer/prompts';
 import {SoloError} from '../../core/errors.js';
 import {type NamespaceName} from '../../core/kube/resources/namespace/namespace_name.js';
-import {type ClusterRef, type DeploymentName, type EmailAddress} from '../../core/config/remote/types.js';
 import {inject, injectable} from 'tsyringe-neo';
 import {InjectTokens} from '../../core/dependency_injection/inject_tokens.js';
 import {type ConfigManager} from '../../core/config_manager.js';
@@ -18,6 +17,10 @@ import {patchInject} from '../../core/dependency_injection/container_helper.js';
 import {ErrorMessages} from '../../core/error_messages.js';
 import {type K8Factory} from '../../core/kube/k8_factory.js';
 import {type LocalConfig} from '../../core/config/local_config.js';
+import {type ClusterResetConfigClass} from './config_interfaces/cluster_reset_config_class.js';
+import {type ClusterSetupConfigClass} from './config_interfaces/cluster_setup_config_class.js';
+import {type ClusterRefConnectConfigClass} from './config_interfaces/cluster_ref_connect_config_class.js';
+import {type ClusterRefDefaultConfigClass} from './config_interfaces/cluster_ref_default_config_class.js';
 
 export const CONNECT_CONFIGS_NAME = 'connectConfig';
 export const DEFAULT_CONFIGS_NAME = 'defaultConfig';
@@ -124,48 +127,4 @@ export class ClusterCommandConfigs {
 
     return ctx.config;
   }
-}
-
-export interface ClusterRefConnectConfigClass {
-  cacheDir: string;
-  devMode: boolean;
-  quiet: boolean;
-  userEmailAddress: EmailAddress;
-  clusterRef: ClusterRef;
-  context: string;
-}
-
-export interface ClusterRefConnectContext {
-  config: ClusterRefConnectConfigClass;
-}
-
-export interface ClusterRefDefaultConfigClass {
-  cacheDir: string;
-  devMode: boolean;
-  clusterRef: string;
-}
-
-export interface ClusterSetupConfigClass {
-  chartDir: string;
-  clusterSetupNamespace: NamespaceName;
-  deployMinio: boolean;
-  deployPrometheusStack: boolean;
-  soloChartVersion: string;
-}
-
-export interface ClusterResetConfigClass {
-  clusterName: string;
-  clusterSetupNamespace: string;
-}
-
-export interface SelectClusterContextContext {
-  config: {
-    quiet: boolean;
-    namespace: NamespaceName;
-    clusterName: string;
-    context: string;
-    clusters: string[];
-    deployment: DeploymentName;
-    deploymentClusters: string[];
-  };
 }
